% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{combine_data}
\alias{combine_data}
\title{Combine Data Into One Object}
\usage{
combine_data(
  taxonomy,
  url_list,
  names,
  species_names,
  subspecies_names,
  strain_ID,
  sentences_fermentation,
  sentences_fermentation_to_acetate
)
}
\arguments{
\item{taxonomy}{A dataframe with columns for taxonomic ranks and rows for articles}

\item{url_list}{A character vector of url names for articles}

\item{names}{A list of organism names (one element per organism)}

\item{species_names}{A list of species names (one element per organism)}

\item{subspecies_names}{A list of subspecies names (one element per organism)}

\item{strain_ID}{A list of type strain IDs (one element per organism)}

\item{sentences_fermentation}{A list of sentences related to fermentation (one element per organism)}

\item{sentences_fermentation_to_acetate}{A list of sentences related to acetate production during fermentation (one element per organism)}
}
\value{
A dataframe with columns for taxonomy and other information and rows for each organism
}
\description{
This function combines information about each organism into one dataframe
}
