% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{create_evaluation_file}
\alias{create_evaluation_file}
\title{Create CSV for Evaluation Metrics}
\usage{
create_evaluation_file(parameters, evaluation_fp)
}
\arguments{
\item{parameters}{A dataframe containing the parameters}
}
\value{
A dataframe with columns for evalaution metrics
}
\description{
This function creates a csv file where evaluation metrics can be stored
}
