% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_articles.R
\name{download_html_pages}
\alias{download_html_pages}
\title{Download HTML Pages}
\usage{
download_html_pages(directory_fp, url_list, destfile_names_Bergey)
}
\arguments{
\item{directory_fp}{A character string giving the file path for the download directory}

\item{url_list}{A character vector of url names for articles}

\item{destfile_names_Bergey}{A list of file names for articles}
}
\description{
This function downloads articles from Bergey's Manual
}
