% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{evaluate_predictions}
\alias{evaluate_predictions}
\title{Evaluate Predictions}
\usage{
evaluate_predictions(y_pred, y_eval)
}
\arguments{
\item{y_pred}{A numeric vector of predictions (1=trait positive, 0=trait negative)}

\item{y_eval}{A numeric vector of labels used as evaluation data}
}
\value{
A list containing numeric scalars for accuracy, precision, and sensitivity
}
\description{
This function calculates accuracy, precision, and sensivity of predictions
}
