% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_articles.R
\name{filter_url_names}
\alias{filter_url_names}
\title{Filter Url Names For Articles}
\usage{
filter_url_names(url_list)
}
\arguments{
\item{url_list}{A character vector of url names for articles}
}
\value{
A character vector of url names for articles
}
\description{
This function keeps url name only if article is for a genus (not phylum, class, order, or family)
}
