% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_full_text}
\alias{get_full_text}
\title{Get Full Text}
\usage{
get_full_text(
  directory_fp,
  url_list,
  destfile_names_Bergey,
  html_list = html_list
)
}
\arguments{
\item{directory_fp}{A character string giving the file path for the download directory}

\item{url_list}{A character vector of url names for articles}

\item{destfile_names_Bergey}{A list of file names for articles}

\item{html_list}{A xml_document and xml_node object for the articles}
}
\value{
A list containing the full text (one element per article)
}
\description{
This function gets the full text from the downloaded articles
}
