% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_genus_text}
\alias{get_genus_text}
\title{Get Genus Text}
\usage{
get_genus_text(url_list, text_full, text_tables)
}
\arguments{
\item{url_list}{A character vector of url names for articles}

\item{text_full}{A list the full text (one element per article)}

\item{text_tables}{A list containing the text of tables (one element per article)}
}
\value{
A list containing the genus text (one element per article)
}
\description{
This function gets text pertaining to the genus (all text between abstract and before species descriptions)
}
