% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{get_ID}
\alias{get_ID}
\title{Get IDs}
\usage{
get_ID(raw_data, ID_colname)
}
\arguments{
\item{raw_data}{A dataframe containing the raw data (columns for ID, genus, species, text, and labels)}

\item{ID_colname}{A character string giving the column name for IDs in the raw data}
}
\value{
A numeric vector IDs with one element per species
}
\description{
This function gets species IDs from the raw data
}
