% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{get_labels}
\alias{get_labels}
\title{Get Labels}
\usage{
get_labels(raw_data, labels_colname)
}
\arguments{
\item{raw_data}{A dataframe containing the data}

\item{labels_colname}{A character string giving the column name for labels in the raw data}
}
\value{
A numeric vector of data labels with one element per species (1=trait positive, 0=trait negative)
}
\description{
This function gets data labels from the raw data and formats it for input into the neural network
}
