% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_main_text}
\alias{get_main_text}
\title{Get Main Text}
\usage{
get_main_text(url_list, text_full)
}
\arguments{
\item{url_list}{A character vector of url names for articles}

\item{text_full}{A list the full text (one element per article)}
}
\value{
A list containing the main text (one element per article)
}
\description{
This function gets all text between abstract and references (or acknowledgements, end notes, further reading)
}
