% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_organisms.R
\name{get_organism_names}
\alias{get_organism_names}
\title{Get Organism Names}
\usage{
get_organism_names(url_list, taxonomy, text_species)
}
\arguments{
\item{url_list}{A character vector of url names for articles}

\item{taxonomy}{A dataframe with columns for taxonomic ranks and rows for articles}

\item{text_species}{A list containing the species text (one element per article)}
}
\value{
A list of organism names (one element per organism)
}
\description{
This function gets the genus, species, and subspecies names of each organism
}
