% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentences_for_traits.R
\name{get_sentences_fermentation}
\alias{get_sentences_fermentation}
\title{Get Sentences Related to Fermentation}
\usage{
get_sentences_fermentation(url_list, names, text_genus, species_description)
}
\arguments{
\item{url_list}{A character vector of url names for articles}

\item{names}{A list of organism names (one element per organism)}

\item{text_genus}{A list containing the genus text (one element per organism)}

\item{species_description}{A list containing the species descriptions (one element per organism)}
}
\value{
A list of sentences related to fermentation (one element per organism)
}
\description{
This function finds sentences in the organism description including keywords related to fermentation
}
