% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_organisms.R
\name{get_species_description}
\alias{get_species_description}
\title{Get Species Description}
\usage{
get_species_description(url_list, names, text_species)
}
\arguments{
\item{url_list}{A character vector of url names for articles}

\item{names}{A list of organism names (one element per organism)}

\item{text_species}{A list containing the species text (one element per article)}
}
\value{
A list containing the species descriptions (one element per organism)
}
\description{
This function gets a description of each organism
It finds the part of the species text that is relevant to each organism
}
