% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_organisms.R
\name{get_species_names}
\alias{get_species_names}
\title{Get Species Names}
\usage{
get_species_names(url_list, taxonomy, names)
}
\arguments{
\item{url_list}{A character vector of url names for articles}

\item{taxonomy}{A dataframe with columns for taxonomic ranks and rows for articles}

\item{names}{A list of organism names (one element per organism)}
}
\value{
A list of species names (one element per organism)
}
\description{
This function gets species names for each organism
}
