% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_species_text}
\alias{get_species_text}
\title{Get Species Text}
\usage{
get_species_text(url_list, text_full, text_tables)
}
\arguments{
\item{url_list}{A character vector of url names for articles}

\item{text_full}{A list the full text (one element per article)}

\item{text_tables}{A list containing the text of tables (one element per article)}
}
\value{
A list containing the species text (one element per article)
}
\description{
This function gets text under "List of the species of the genus" or similar section
}
