% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_organisms.R
\name{get_strain_ID}
\alias{get_strain_ID}
\title{Get Strain ID}
\usage{
get_strain_ID(url_list, names, species_description)
}
\arguments{
\item{url_list}{A character vector of url names for articles}

\item{names}{A list of organism names (one element per organism)}

\item{species_description}{A list containing the species descriptions (one element per organism)}
}
\value{
A list of type strain IDs (one element per organism)
}
\description{
This function gets type the strain ID for each organism
}
