% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_organisms.R
\name{get_subspecies_names}
\alias{get_subspecies_names}
\title{Get Subspecies Names}
\usage{
get_subspecies_names(url_list, names)
}
\arguments{
\item{url_list}{A character vector of url names for articles}

\item{names}{A list of organism names (one element per organism)}
}
\value{
A list of subspecies names (one element per organism)
}
\description{
This function gets subspecies names for each organism
}
