% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_tables}
\alias{get_tables}
\title{Get Tables}
\usage{
get_tables(directory_fp, url_list, destfile_names_Bergey, html_list)
}
\arguments{
\item{directory_fp}{A character string giving the file path for the download directory}

\item{url_list}{A character vector of url names for articles}

\item{destfile_names_Bergey}{A list of file names for articles}

\item{html_list}{A xml_document and xml_node object for the articles}
}
\value{
A list containing the text of tables (one element per article)
}
\description{
This function gets text of tables from the downloaded articles
}
