% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{get_text}
\alias{get_text}
\title{Get Text}
\usage{
get_text(raw_data, text_colname)
}
\arguments{
\item{raw_data}{A dataframe containing the data}

\item{labels_colname}{A character string giving the column name for text in the raw data}
}
\value{
A character vector of text with one element per species
}
\description{
This function gets text from the raw data and formats it for tokenization
}
