% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_articles.R
\name{get_url_names}
\alias{get_url_names}
\title{Get Url Names For Articles}
\usage{
get_url_names()
}
\value{
A character vector of url names for articles
}
\description{
This function gets url names for articles for articles from Bergey's Manual table of contents
}
