% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{load_parameters}
\alias{load_parameters}
\title{Load Parameters}
\usage{
load_parameters(parameters_fp)
}
\arguments{
\item{parameters_fp}{A character string giving the file path of the csv file}
}
\value{
A dataframe containing the parameters
}
\description{
This function loads a csv file containing parameters to run the neural network
}
