% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{make_predictions}
\alias{make_predictions}
\title{Make Predictions}
\usage{
make_predictions(model_fp, x_eval)
}
\arguments{
\item{model_fp}{The file path where the model is saved in HDF5 format}

\item{x_eval}{A matrix of tokenized and paddded text used as evaluation data}
}
\value{
A numeric vector of predictions (1=trait positive, 0=trait negative)
}
\description{
This function loads a trained neural network model (HDF5 format) then makes predictions
}
