% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{pad_text}
\alias{pad_text}
\title{Pad Text}
\usage{
pad_text(text_seqs, maxlen = 3000)
}
\arguments{
\item{text_seqs}{A list of tokenized text with one element per species}

\item{maxlen}{A numeric scalar giving the length of text (in words); additional words are truncated}
}
\value{
A matrix of tokenized and padded text
}
\description{
This function pads the tokenized text for input into the network
}
