% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{read_html_pages}
\alias{read_html_pages}
\title{Read HTML pages}
\usage{
read_html_pages(directory_fp, url_list, destfile_names_Bergey)
}
\arguments{
\item{directory_fp}{A character string giving the file path for the download directory}

\item{url_list}{A character vector of url names for articles}

\item{destfile_names_Bergey}{A list of file names for articles (one element per page)}
}
\value{
A xml_document and xml_node object for the articles
}
\description{
This function retrieves downloaded articles and reads them into R
}
