% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{save_evaluation}
\alias{save_evaluation}
\title{Save Evaluation Metrics}
\usage{
save_evaluation(
  accuracy,
  precision,
  sensitivity,
  evaluation_number,
  evaluation_fp
)
}
\arguments{
\item{accuracy}{A numeric scalar of the accuracy of predictions}

\item{precision}{A numeric scalar of the precision of predictions}

\item{sensitivity}{A numeric scalar of the sensitivity of predictions}

\item{evaluation_number}{A numeric scalar for
#' @export}

\item{evaluation_fp}{A character string giving the file path of the csv file}
}
\description{
This function writes evaluation metrics to a csv file
}
