% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{set_data}
\alias{set_data}
\title{Set Data}
\usage{
set_data(text_seqs_padded, labels, train_fraction)
}
\arguments{
\item{text_seqs_padded}{A matrix of tokenized and padded text}

\item{labels}{A numeric vector of data labels with one element per species (1=positive, 0=negative)}

\item{train_fraction}{A numeric scalar giving the fraction of data to be used for training}
}
\value{
A list of training and evaluation data
}
\description{
This function splits labels and text into training and evaluation data for input into the network
}
