% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_articles.R
\name{set_html_names}
\alias{set_html_names}
\title{Set HTML Names}
\usage{
set_html_names(url_list)
}
\arguments{
\item{url_list}{A character vector of url names for articles}
}
\value{
A list of file names for articles
}
\description{
This function sets file names for articles when downloaded
}
