% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{set_model}
\alias{set_model}
\title{Set Model Architecture}
\usage{
set_model(
  maxlen = 3000,
  max_vocab = 3000,
  embedding_dim = 50,
  filters = 64,
  kernel_size = 3,
  hidden_dim = 50
)
}
\arguments{
\item{maxlen}{A numeric scalar giving the length of text (in words); additional words are truncated}

\item{max_vocab}{A numeric scalar giving the number of unique words to keep in the text; additional words are removed}

\item{embedding_dim}{A numeric scalar of the output dimensions for the embedding layer}

\item{filters}{A numeric scalar of the number of fulters for convolution}

\item{kernel_size}{A numeric scalar of the kernel size for convolution}

\item{hidden_dim}{A numeric scalar of the output dimensions for the hidden dense layer}
}
\value{
A Keras neural network model
}
\description{
This function defines the architecture of the neural network model in Keras
}
