% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{set_model_RNN}
\alias{set_model_RNN}
\title{Set Model Architecture for RNN}
\usage{
set_model_RNN(
  maxlen = 20000,
  max_vocab = 3000,
  embedding_dim = 128,
  lstm_units = 64
)
}
\arguments{
\item{maxlen}{A numeric scalar giving the length of text (in words); additional words are truncated}

\item{max_vocab}{A numeric scalar giving the number of unique words to keep in the text; additional words are removed}

\item{embedding_dim}{A numeric scalar of the output dimensions for the embedding layer}

\item{lstm_units}{A numeric scalar of the number LSTM units}
}
\value{
A Keras neural network model
}
\description{
This function defines the architecture of the neural network model in Keras
}
