% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{tokenize_text}
\alias{tokenize_text}
\title{Tokenize Text}
\usage{
tokenize_text(text, nchar_max = 25000, max_vocab = 3000)
}
\arguments{
\item{text}{A character vector of text with one element per species}

\item{nchar_max}{A numeric scalar giving the number of unique words to keep in the text; additional words are removed}
}
\value{
A list of tokenized text with one element per species
}
\description{
This function converts text (written species descriptions) into machine-readable tokens
}
