% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{train_model}
\alias{train_model}
\title{Train Model}
\usage{
train_model(model, x_train, y_train, batch_size = 32, epochs = 10, model_fp)
}
\arguments{
\item{model}{A Keras neural network model}

\item{x_train}{A matrix of of tokenized and paddded text used as training data}

\item{y_train}{A numeric vector of labels used as training data}

\item{batch_size}{A numeric scalar of the batch size used in training}

\item{epochs}{A numeric scalar of the number of epochs used for training}

\item{model_fp}{The file path to save the model in HDF5 format}
}
\value{
A list of tokenized text with one element per species
}
\description{
This function trains the neural network model and saves it in HDF5 format
}
